Reorganize ui/theme into separate folder
