+++ packages/ui/src/components/Badge/Badge.helper.ts
@@ -1,6 +1,6 @@
 import { CSSObject } from '@emotion/react'
-import { Level, mq } from '../../lib/media-query'
-import { theme } from '../../lib/theme/theme'
+import { theme } from '../../theme'
+import { Level, mq } from '../../theme/media-query'
 
 type PartialRecord<K extends keyof any, T> = Partial<Record<K, T>>
