+++ packages/ui/src/components/Badge/Badge.tsx
@@ -1,8 +1,8 @@
 import isPropValid from '@emotion/is-prop-valid'
 import styled from '@emotion/styled'
-import { getMargins, Margins } from '../../lib/margins'
-import { UIColors } from '../../lib/theme/colors/colors'
-import { getColor, theme } from '../../lib/theme/theme'
+import { getColor, theme } from '../../theme'
+import { UIColors } from '../../theme/colors/colors'
+import { getMargins, Margins } from '../../theme/margins'
 import { BadgeSizeProps, getBadgeSize } from './Badge.helper'
 
 type BadgeColors = Pick<
