+++ packages/ui/src/components/Heading/Heading.helpers.ts
@@ -1,6 +1,6 @@
 import { CSSObject } from '@emotion/react'
-import { mq, Level } from '../../lib/media-query'
-import { theme } from '../../lib/theme/theme'
+import { theme } from '../../theme'
+import { Level, mq } from '../../theme/media-query'
 
 type StandardHeadingSize = '18' | '20' | '24' | '32' | '40' | '48' | '56' | '72' | '96'
 type SerifHeadingSize = Exclude<StandardHeadingSize, '18'>
