+++ packages/ui/src/components/Heading/Heading.tsx
@@ -2,9 +2,9 @@ import isPropValid from '@emotion/is-prop-valid'
 import styled from '@emotion/styled'
 import React from 'react'
 import Balancer from 'react-wrap-balancer'
-import { getMargins, Margins } from '../../lib/margins'
-import { UIColors } from '../../lib/theme/colors/colors'
-import { getColor, theme } from '../../lib/theme/theme'
+import { getColor, theme } from '../../theme'
+import { UIColors } from '../../theme/colors/colors'
+import { getMargins, Margins } from '../../theme/margins'
 import { getHeadingVariantStyles, HeadingVariant } from './Heading.helpers'
 
 export type { PossibleHeadingVariant } from './Heading.helpers'
