+++ packages/ui/src/components/HedvigLogo/HedvigLogo.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryFn } from '@storybook/react'
-import { theme } from '../..'
+import { theme } from '../../theme'
 import { HedvigLogo } from './HedvigLogo'
 
 export default {
