+++ packages/ui/src/components/HedvigSymbol/HedvigSymbol.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryFn } from '@storybook/react'
-import { theme } from '../..'
+import { theme } from '../../theme'
 import { HedvigSymbol } from './HedvigSymbol'
 
 export default {
