+++ packages/ui/src/theme/typography.ts
@@ -1,3 +1,6 @@
+import type { CSSObject } from '@emotion/react'
+import { Level, mq } from './media-query'
+
 export enum HedvigFont {
   HEDVIG_LETTERS_BIG = 'HedvigLetters-Big',
   HEDVIG_LETTERS_SMALL = 'HedvigLetters-Small',
@@ -43,3 +46,31 @@ export const fontSizes = {
 } as const
 
 export type FontSizes = keyof typeof fontSizes
+
+export type PartialRecord<K extends keyof any, T> = Partial<Record<K, T>>
+
+export type FontSizeProps = FontSizes | PartialRecord<Level | '_', FontSizes>
+
+export const getFontSize = (sizes: FontSizeProps): CSSObject => {
+  if (typeof sizes !== 'object') {
+    return { fontSize: fontSizes[sizes] }
+  }
+  const styles = {} as CSSObject
+
+  const breakpoints = Object.keys(sizes) as Array<keyof typeof sizes>
+  breakpoints.forEach((breakpoint) => {
+    const sizeAtBreakpoint = sizes[breakpoint]
+    if (!sizeAtBreakpoint) {
+      return
+    }
+    if (breakpoint === '_') {
+      // Default
+      styles.fontSize = fontSizes[sizeAtBreakpoint]
+    } else {
+      styles[mq[breakpoint]] = {
+        fontSize: fontSizes[sizeAtBreakpoint],
+      }
+    }
+  })
+  return styles
+}
